+++ bignum/src/commonMain/kotlin/com/ionspin/kotlin/bignum/decimal/BigDecimal.kt
@@ -35,7 +35,7 @@ import kotlin.math.min
  * Implementation of floating-point arbitrary precision arithmetic.
  *
  * Each object of this class represents an immutable large floating point number. The underlying implementation
- * uses [BigInteger] to represent sign and significand, and another [BigInteger] to represent the exponent.
+ * uses [BigInteger] to represent sign and significand, and [Long] to represent the exponent.
  *
  * [DecimalMode] defines the precision and desired [RoundingMode] when using instances of this class in arithmetic operations,
  * but if [DecimalMode] is supplied to the operation it will override instances [DecimalMode]
@@ -46,13 +46,13 @@ import kotlin.math.min
  *
  * When arithmetic operations have both operands unlimited precision and no scaling, the result is
  * also unlimited precision and no scale. When an operation mixes an unlimited precision operand
- * and a scaled operand, the result is unlimited precision. WHen both operands have scale,
+ * and a scaled operand, the result is unlimited precision. When both operands have scale,
  * whether unlimited precision or limited precision, then these rules for scale of the result are used:
  * <ul>
  *     <li>add, subtract - max of the two scales</li>
  *     <li>multiply - sum of the two scales</li>
  *     <li>divide - min of the two scales</li>
- *
+ * </ul>
  * Created by Ugljesa Jovanovic
  * ugljesa.jovanovic@ionspin.com
  * on 23-Mar-2019
@@ -293,33 +293,14 @@ class BigDecimal private constructor(
                 RoundingMode.ROUND_HALF_TO_EVEN -> {
                     when {
                         decider == SignificantDecider.FIVE -> {
-                            if (significand % 2 == BigInteger.ONE) {
+                            if ((significand % 2).abs() == BigInteger.ONE) {
                                 // RoundingMode.HALF_CEILING if the digit to the left of the discarded fraction is odd
                                 when (sign) {
                                     Sign.POSITIVE -> {
-                                        if (decider != SignificantDecider.LESS_THAN_FIVE) {
-                                            result++
-                                        }
+                                        result++
                                     }
                                     Sign.NEGATIVE -> {
-                                        if (decider != SignificantDecider.LESS_THAN_FIVE) {
-                                            result--
-                                        }
-                                    }
-                                    Sign.ZERO -> {
-                                    }
-                                }
-                            } else {
-                                when (sign) {
-                                    Sign.POSITIVE -> {
-                                        if (decider == SignificantDecider.MORE_THAN_FIVE) {
-                                            result++
-                                        }
-                                    }
-                                    Sign.NEGATIVE -> {
-                                        if (decider != SignificantDecider.LESS_THAN_FIVE) {
-                                            result--
-                                        }
+                                        result--
                                     }
                                     Sign.ZERO -> {
                                     }
@@ -343,29 +324,10 @@ class BigDecimal private constructor(
                                 // RoundingMode.HALF_CEILING if the digit to the left of the discarded fraction is even
                                 when (sign) {
                                     Sign.POSITIVE -> {
-                                        if (decider != SignificantDecider.LESS_THAN_FIVE) {
-                                            result++
-                                        }
+                                        result++
                                     }
                                     Sign.NEGATIVE -> {
-                                        if (decider != SignificantDecider.LESS_THAN_FIVE) {
-                                            result--
-                                        }
-                                    }
-                                    Sign.ZERO -> {
-                                    }
-                                }
-                            } else {
-                                when (sign) {
-                                    Sign.POSITIVE -> {
-                                        if (decider == SignificantDecider.MORE_THAN_FIVE) {
-                                            result++
-                                        }
-                                    }
-                                    Sign.NEGATIVE -> {
-                                        if (decider != SignificantDecider.LESS_THAN_FIVE) {
-                                            result--
-                                        }
+                                        result--
                                     }
                                     Sign.ZERO -> {
                                     }
@@ -488,8 +450,13 @@ class BigDecimal private constructor(
             return if (exponent >= 0) {
                 roundSignificand(significand, exponent, workMode)
             } else {
-                val temp = BigDecimal(significand, exponent) + significand.signum()
-                roundSignificand(temp.significand, temp.exponent, workMode) - significand.signum()
+                if (decimalMode.roundingMode == RoundingMode.ROUND_HALF_TO_EVEN) {
+                    val temp = BigDecimal(significand, exponent) + (significand.signum() * 2)
+                    roundSignificand(temp.significand, temp.exponent, workMode) - (significand.signum() * 2)
+                } else {
+                    val temp = BigDecimal(significand, exponent) + significand.signum()
+                    roundSignificand(temp.significand, temp.exponent, workMode) - significand.signum()
+                }
             }
         }
 
@@ -897,7 +864,7 @@ class BigDecimal private constructor(
          */
         fun parseStringWithMode(floatingPointString: String, decimalMode: DecimalMode? = null): BigDecimal {
             if (floatingPointString.isEmpty()) {
-                return ZERO
+                throw ArithmeticException("Empty string is not a valid decimal number")
             }
             if (floatingPointString.contains('E', true)) {
                 // Sci notation
@@ -1359,7 +1326,7 @@ class BigDecimal private constructor(
             return Pair(ZERO, this)
         }
         val resolvedRoundingMode =
-            this.decimalMode?.copy(decimalPrecision = exponent + 1) ?: DecimalMode(exponent + 1, RoundingMode.FLOOR)
+            this.decimalMode?.copy(decimalPrecision = exponent - other.exponent + 1) ?: DecimalMode(exponent - other.exponent + 1, RoundingMode.FLOOR)
         val quotient = divide(other, resolvedRoundingMode)
         val quotientInfinitePrecision = quotient.copy(decimalMode = DecimalMode.DEFAULT)
         val remainder = this - (quotientInfinitePrecision * other)
@@ -1469,7 +1436,6 @@ class BigDecimal private constructor(
                 Sign.NEGATIVE -> ONE.negate()
                 Sign.ZERO -> ZERO
             }
-            return ZERO
         }
         return roundSignificand(DecimalMode(exponent + 1, RoundingMode.FLOOR))
     }
@@ -1870,7 +1836,7 @@ class BigDecimal private constructor(
                     }
                     bitList.add(bit)
                     fractionConvertTemp = if (bit == 1) {
-                        (multiplied divrem TEN).second
+                        (multiplied divrem ONE).second
                     } else {
                         multiplied
                     }
@@ -1925,7 +1891,7 @@ class BigDecimal private constructor(
             }
             // For significand we can have a maximum of 53 (52 + 1 implicit)
             // If there is no decimal point at all we can directly count the bits in significand and use them,
-            // but if there is we need to convert the fractional part to binary32 representation first.
+            // but if there is we need to convert the fractional part to binary64 representation first.
             // Bit count:
             val totalBits = if (precision - exponent - 1 > 0) {
                 // First find out where the decimal point will be
@@ -1950,7 +1916,7 @@ class BigDecimal private constructor(
                     }
                     bitList.add(bit)
                     fractionConvertTemp = if (bit == 1) {
-                        (multiplied divrem TEN).second
+                        (multiplied divrem ONE).second
                     } else {
                         multiplied
                     }
@@ -2025,7 +1991,11 @@ class BigDecimal private constructor(
         val rounded = if (this.exponent >= 0) {
             roundSignificand(DecimalMode(digitPosition, roundingMode))
         } else {
-            (this + this.signum()).roundSignificand(DecimalMode(digitPosition, roundingMode)) - this.signum()
+            if (roundingMode == RoundingMode.ROUND_HALF_TO_EVEN) {
+                (this + this.signum() * 2).roundSignificand(DecimalMode(digitPosition, roundingMode)) - this.signum() * 2
+            } else {
+                (this + this.signum()).roundSignificand(DecimalMode(digitPosition, roundingMode)) - this.signum()
+            }
         }
 
         return if (decimalMode == null) {
@@ -2139,7 +2109,7 @@ class BigDecimal private constructor(
 
     override fun compareTo(other: Any): Int {
         if (other is Number) {
-            if (RuntimePlatform.currentPlatform() == Platform.JS) {
+            if (RuntimePlatform.currentPlatform() == Platform.JS || RuntimePlatform.currentPlatform() == Platform.WASMJS) {
                 return javascriptNumberComparison(other)
             }
         }