+++ bignum/src/commonTest/kotlin/com/ionspin/kotlin/bignum/decimal/ReportedIssueReplicationTest.kt
@@ -332,4 +332,24 @@ class ReportedIssueReplicationTest {
         println("Is using scale? ${bigDecimal.usingScale} Scale ${bigDecimal.scale}")
         println("DecimalMode ${bigDecimal.decimalMode}")
     }
+
+    @Test
+    fun testRemainder() {
+        val a = BigDecimal.fromDouble(1815.5)
+        val b = BigDecimal.fromInt(360)
+
+        val (q, r) = a.divideAndRemainder(b)
+        assertEquals("5", q.toStringExpanded())
+        assertEquals("15.5", r.toStringExpanded())
+    }
+
+    @Test
+    fun roundingHalfEven() {
+        val bigDecimal = BigDecimal.parseString("0.5")
+        val rounded = bigDecimal.roundToDigitPositionAfterDecimalPoint(0, roundingMode = RoundingMode.ROUND_HALF_TO_EVEN)
+        assertEquals(
+            BigDecimal.ZERO,
+            rounded
+        )
+    }
 }